Merged PR 1524: 203,204,205 can be executed successfully
Once linked with devops are able to demo runnable, automated UI testing.

As per 204 "All selected files appear in the list of files, the filenames are displayed, the upload time is displayed in UTC format, the upload of the files happen in the background after the user saves the changes"

The Case should be re-opened after processing complete and the upload time checked in the case details. The test case is old and doesnt reflect the current system.

Related work items: #6147